requirements.txt
Django
django_ipware
django_tables2
django-bootstrap-datepicker
django_crispy_forms
django-sequences
google-cloud-logging
django_model_utils
Pillow

django-filter
django-widget-tweaks
django-bootstrap-datepicker-plus

bs4
geopy

django-hashid-field
psycopg2-binary
django-autocomplete-light
xlrd
xlwt
django-sequences
gunicorn

zeep